Enable fine-tuning on custom (ELEVANT) datasets via cmdl args
--training_file and --evaluation_file can be specified in order to
fine-tune and/or evaluate on custom datasets in the ELEVANT JSONL
format. Entity Mentions from the JSONL format are currently used as
ground truth labels. In order to use benchmark labels from the JSONL
format, the Oracle linker must first be run over the corresponding
benchmark and the resulting file can then be used as training or
evaluation file.